requirements.txt
asgiref==3.8.1
Django==4.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
inflection==0.5.1
mssql-django==1.5
packaging==24.1
PyJWT==2.9.0
pyodbc==5.2.0
pytz==2024.2
PyYAML==6.0.2
setuptools==75.1.0
sqlparse==0.5.1
tzdata==2024.2
uritemplate==4.1.1
wheel==0.44.0
